serv_cda_manager_crud/requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2022.7.1
typing-extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
certifi
flask_cors==4.0.0